Execute workaround as setup stop on sbuild
Default configuration has eatmydata configured, so schroot commands
don't write anything to disk between different commands.

When sbuild runs, it creates a session internally between the different
schroot commands it executes. Then we can use one of the steps
(chroot-setup-commands) to compile the workaround and load it to
ld.so.preload.

Also making the workaround non-conditional for now - since we are only
building for arm at the moment.

Signed-off-by: Daniel Finimundi <daniel@finimundi.com>