Workaround for bug on qemu 5.2
In order to fix the segmentation fault on qemu 4.2
(https://bugs.launchpad.net/qemu/+bug/1749393), there is a fix on
version >=5.2.

Although, this also introduced another bug with current version of
schroot
(https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=983087). This is
fixed on 1.6.10-12, but that is not available on Ubuntu 20.04 as well.
Fortunatelly this one has a simple workaround, which is to create the
missing directory before touch'ing the file (which is done when entering
the schroot).

Signed-off-by: Daniel Finimundi <daniel@finimundi.com>